add func unit behavior

+++ src/sim/register.rs
@@ -1,4 +1,6 @@
-pub type RegisterIdType = usize;
+pub type RegisterIdType = u32;
+
+#[derive(Debug, Clone, PartialEq, Eq)]
 pub enum RegisterType {
     ScalarRegister(RegisterIdType),
     VectorRegister(RegisterIdType),
